[refactor] GreetingComponent 위치 변경
